Fixed some organiztion
