+++ Housing.Data/Housing.Data.Domain/CRUD/Deletions.cs
@@ -1,0 +1,56 @@
+﻿using Housing.Data.Domain.DataAccessObjects;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Housing.Data.Domain.CRUD
+{
+    public partial class AccessHelper
+    {
+        #region deletions 
+
+        /// <summary>
+        /// delete Associate
+        /// </summary>
+        /// <param name="assoc"></param>
+        /// <returns>true if deletion successful</returns>
+        public bool DeleteAssociate(AssociateDao assoc)
+        {
+            return true;
+        }
+
+        /// <summary>
+        /// delete HousingComplex 
+        /// </summary>
+        /// <param name="hc"></param>
+        /// <returns>true if deletion successful</returns>
+        public bool DeleteHousingComplex(HousingComplexDao hc)
+        {
+            return true;
+        }
+
+        /// <summary>
+        /// delete HousingUnit
+        /// </summary>
+        /// <param name="hu"></param>
+        /// <returns>true if deletion successful</returns>
+        public bool DeleteHousingUnit(HousingUnitDao hu)
+        {
+            return true;
+        }
+
+        /// <summary>
+        /// delete HousingData entry
+        /// </summary>
+        /// <param name="hd"></param>
+        /// <returns>true if deletion successful</returns>
+        public bool DeleteHousingData(HousingDataDao hd)
+        {
+            return true;
+        }
+
+        #endregion
+    }
+}
