+++ Housing.Data/Housing.Data.Domain/CRUD/Updates.cs
@@ -1,0 +1,57 @@
+﻿using Housing.Data.Domain.DataAccessObjects;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Housing.Data.Domain.CRUD
+{
+    public partial class AccessHelper
+    {
+
+        #region updates
+
+        /// <summary>
+        /// update Associate
+        /// </summary>
+        /// <param name="assoc"></param>
+        /// <returns>true if update successful</returns>
+        public bool UpdateAssociate(AssociateDao assoc)
+        {
+            return true;
+        }
+
+        /// <summary>
+        /// update HousingComplex 
+        /// </summary>
+        /// <param name="hc"></param>
+        /// <returns>true if update successful</returns>
+        public bool UpdateHousingComplex(HousingComplexDao hc)
+        {
+            return true;
+        }
+
+        /// <summary>
+        /// update HousingUnit
+        /// </summary>
+        /// <param name="hu"></param>
+        /// <returns>true if update successful</returns>
+        public bool UpdateHousingUnit(HousingUnitDao hu)
+        {
+            return true;
+        }
+
+        /// <summary>
+        /// update HousingData entry
+        /// </summary>
+        /// <param name="hd"></param>
+        /// <returns>true if update successful</returns>
+        public bool UpdateHousingData(HousingDataDao hd)
+        {
+            return true;
+        }
+
+        #endregion
+    }
+}
